% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_bplot.R
\name{ctd_bplot}
\alias{ctd_bplot}
\title{Plot contours of bottom water variable by station and time}
\usage{
ctd_bplot(dat_in, var_plo, date_col = "Date", num_levs = 8, expand = 200,
  txt_scl = 1, xlab = "Date", ylab = "Channel distance (km)",
  cols = c("tomato", "lightblue", "lightgreen", "green"), ncol = 100)
}
\arguments{
\item{dat_in}{ctd data along the tidal axis, as a \code{\link[base]{data.frame}}}

\item{var_plo}{chr string of variable to plot from \code{dat_in}}

\item{date_col}{chr string of name of date column}

\item{num_levs}{numeric for number of contour levels}

\item{expand}{numeric for expanding sample size of CTD points, see \code{\link{get_depths}}}

\item{txt_scl}{numeric for scaling all text labels}

\item{xlab}{chr string for x-axis label}

\item{ylab}{chr string for y-axis label}

\item{cols}{chr string of colors for the plot}

\item{ncol}{numeric indicating degree of smoothing for the color palette}
}
\value{
The contour plot
}
\description{
Plot two-dimensional CTD contours for bottom water variable by station and time
}
\details{
Raw data from the CTD are vertical profiles at unique stations for more than one date.  The data are linearly interpolated between dates and stations to create a two-dimensional plotting surface. The \code{dat_in} data.frame must have at a minimum six columns for date (\code{Date}), longitude (\code{Long}), latitude (\code{Lat}), station name (\code{Station}), depth (\code{Depth}, non-negative), and the variable to plot.
}
\examples{
# default plot
ctd_bplot(ctd, var_plo = 'DO')
}

