% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_plot.R
\name{ctd_plot}
\alias{ctd_plot}
\title{Contour plot for CTD data}
\usage{
ctd_plot(dat_in, var_plo, dep_in = NULL, date = NULL, date_col = "Date",
  rngs_in = NULL, num_levs = 8, expand = 200, window = 1, chop = 0,
  add = 0, txt_scl = 1, xlab = "Channel distance from P01 to P09 (km)",
  ylab = "Depth (m)", var_lab = NULL, cols = c("tomato", "lightblue",
  "lightgreen", "green"), msk_col = "grey", cont_ext = 0.5, ylim = NULL,
  ncol = 100)
}
\arguments{
\item{dat_in}{ctd data along the tidal axis, as a \code{\link[base]{data.frame}}}

\item{var_plo}{chr string of variable to plot from \code{dat_in}}

\item{dep_in}{depth soundings from bathymetric soundings}

\item{date}{chr string of date to plot, only required if input data has more than one sample date}

\item{date_col}{chr string of name of date column}

\item{rngs_in}{output from \code{\link{get_rngs}}, used to scale the colors, see examples}

\item{num_levs}{numeric for number of contour levels}

\item{expand}{numeric for expanding sample size of CTD points, see \code{\link{get_depths}}}

\item{window}{numeric for smoothing factor to reduce jaggedness of depth values, default does nothing}

\item{chop}{numeric for trimming the depth values}

\item{add}{numeric for scalar to add to all depth values}

\item{txt_scl}{numeric for scaling all text labels}

\item{xlab}{chr string for x-axis label}

\item{ylab}{chr string for y-axis label}

\item{var_lab}{optional chr string of text to include in the plot}

\item{cols}{chr string of colors for the plot}

\item{msk_col}{chr string for color of masking depth}

\item{cont_ext}{numeric (meters) for reducing the mask size for the interpolated contour plot, this can be useful to remove white spaces that are not covered by the masking depth but may have an effect on the gradient in the color ramp}

\item{ylim}{numeric vector for y-axis limits, values must be negative for increasing depth}

\item{ncol}{numeric indicating degree of smoothing for the color palette}
}
\value{
The contour plot
}
\description{
Plot two-dimensional contours for CTD data by depth and distance
}
\details{
Raw data from the CTD are vertical profiles at unique stations.  The data are linearly interpolated along each profile and between stations to create a two-dimensional plotting surface. The \code{dat_in} data.frame must have three columns for longitude (\code{Long}), latitude (\code{Latitude}), and station name (\code{Station}).  The \code{dep_in} data.frame (if provided) should have three columns for longitude (\code{Long}), latitude (\code{Latitude}), and depth (\code{Depth}, non-negative).
}
\examples{
# default plot
ctd_plot(ctd_ex1, var_plo = 'Salinity', dep_in = PB_dep_pts)

# make color ramp match contour categories in legend
ctd_plot(ctd_ex1, 'Salinity', ncol = 8)

# change colors
ctd_plot(ctd_ex1, 'Salinity', cols = c('Blue', 'Purple', 'Orange'))
}
\references{
Hagy III, JD, Murrell, MC. 2007. Susceptibility of a northern Gulf of Mexico estuary to hypoxia: An analysis using box models. Estuarine, Coastal, and Shelf Science. 74:239-253.
}

