% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_plotmult.R
\name{ctd_plotmult}
\alias{ctd_plotmult}
\title{Multiple contour plots for CTD data}
\usage{
ctd_plotmult(dat_in, var_plo, var_labs = NULL, ...)
}
\arguments{
\item{dat_in}{list of input ctd data where each element is a \code{\link[base]{data.frame}} of the vertical profiles for each station on a particular date}

\item{var_plo}{chr string of variable to plot from \code{dat_in}}

\item{var_labs}{chr vector of optional text for each plot}

\item{...}{additional arguments passed to \code{\link{ctd_plot}}}
}
\value{
The contour plots with identical color scaling
}
\description{
Multiple contour plots for CTD data with same color scale
}
\examples{
dat <- list(ctd_ex1, ctd_ex2)

# default plot
ctd_plotmult(dat, 'Salinity', var_labs = c('April', 'May'))
}

