% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/depth_tran.R
\docType{data}
\name{depth_tran}
\alias{depth_tran}
\title{Depth and distance profile for Escambia Bay}
\format{A \code{\link[base]{data.frame}} with 129 rows and 2 variables:
\describe{
  \item{\code{tran_km}}{numeric, distance in km along the axis}
  \item{\code{mllw_m}}{numeric, depth at mean lower low water (meters)}
}}
\usage{
depth_tran
}
\description{
A \code{\link[base]{data.frame}} of distance (km) and depth (m) along the primary longitudinal axis of the Escambia Bay sub-basin of Pensacola Bay, Florida.  Used in \code{\link{ctd_plot}}.
}
\examples{
data(depth_tran)
}
\keyword{datasets}

