% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filled_legend.R
\name{filled_legend}
\alias{filled_legend}
\title{Plot legend from a contour plot}
\usage{
filled_legend(
  x = seq(0, 1, length.out = nrow(z)),
  y = seq(0, 1, length.out = ncol(z)),
  z,
  xlim = range(x, finite = TRUE),
  ylim = range(y, finite = TRUE),
  zlim = range(z, finite = TRUE),
  levels = pretty(zlim, nlevels),
  nlevels = 20,
  color.palette = cm.colors,
  col = color.palette(length(levels) - 1),
  plot.title,
  plot.axes,
  key.title,
  key.axes,
  asp = NA,
  xaxs = "i",
  yaxs = "i",
  las = 1,
  axes = TRUE,
  frame.plot = axes,
  ...
)
}
\arguments{
\item{x}{locations of grid lines at which the values in \code{z} are
   measured.  These must be in ascending order.  (The rest of this
   description does not apply to \code{.filled.contour}.)
   By default, equally spaced values from 0 to 1 are used.  If \code{x}
   is a \code{list}, its components \code{x$x} and \code{x$y} are used
   for \code{x} and \code{y}, respectively.  If the list has component
   \code{z} this is used for \code{z}.}

\item{y}{numeric same as \code{x}}

\item{z}{a numeric matrix containing the values to be plotted..  Note that
   \code{x} can be used instead of \code{z} for convenience.}

\item{xlim}{x limits for the plot.}

\item{ylim}{y limits for the plot.}

\item{zlim}{z limits for the plot.}

\item{levels}{a set of levels which are used to partition the range
    of \code{z}.  Must be \bold{strictly} increasing (and finite).  Areas
    with \code{z} values between consecutive levels are painted with the
    same color.}

\item{nlevels}{if \code{levels} is not specified, the range of \code{z},
    values is divided into approximately this many levels.}

\item{color.palette}{a color palette function to be used to assign
    colors in the plot.}

\item{col}{an explicit set of colors to be used in the plot.
    This argument overrides any palette function specification.  There
    should be one less color than levels}

\item{plot.title}{statements which add titles to the main plot.}

\item{plot.axes}{statements which draw axes (and a \code{\link[graphics]{box}})
   on the main plot.  This overrides the default axes.}

\item{key.title}{statements which add titles for the plot key.}

\item{key.axes}{statements which draw axes on the plot key.
    This overrides the default axis.}

\item{asp}{the \eqn{y/x} aspect ratio, see \code{\link[graphics]{plot.window}}.}

\item{xaxs}{the x axis style.  The default is to use internal
    labeling.}

\item{yaxs}{the y axis style.  The default is to use internal
    labeling.}

\item{las}{the style of labeling to be used.  The default is to
    use horizontal labeling.}

\item{axes}{logicals indicating if axes and a box should be
   drawn, as in \code{\link[graphics]{plot.default}}.}

\item{frame.plot}{logical for frame}

\item{...}{additional \link[graphics]{graphical parameters}, currently only passed to
   \code{\link[graphics]{title}()}.}
}
\description{
Plot only the legend from a contour plot
}
\details{
This function was taken from the filled.legend script at \url{http://wiki.cbr.washington.edu/qerm/index.php/R/Contour_Plots}.  The following was taken directly from original script:

Modification of filled.contour by Carey McGilliard and Bridget Ferris designed to just plot the legend.
}
