% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bdepths.R
\name{get_bdepths}
\alias{get_bdepths}
\title{Interpolate observed CTD bottom depth along sampling axis}
\usage{
get_bdepths(ctd_pts, expand = 200)
}
\arguments{
\item{ctd_pts}{data.frame of CTD station locations, see details}

\item{expand}{numeric for expanding sample size of CTD points}
}
\value{
The expanded data.frame from \code{ctd_pts} including depth (\code{Depth}) and cumulative distance (\code{Dist}, in km) columns.
}
\description{
Interpolate observed CTD bottom depth along sampling axis
}
\details{
CTD station locations and bottom depths are linearly expanded for use in \code{\link{ctd_plot}} if depth soundings are not provided.  This is done to create a smoother depth gradient in the contour plot.  The \code{ctd_pts} data.frame must have three columns for longitude (\code{Long}), latitude (\code{Latitude}), and station name (\code{Station}).

All coordinates are assumed to be geographic decimal degrees using the WGS 1984 projection, negative longitude is west of the Prime Meridian.
}
\examples{
\dontrun{
get_bdepths(ctd_ex1)
}
}

