% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rngs.R
\name{get_rngs}
\alias{get_rngs}
\title{Get variable ranges}
\usage{
get_rngs(dat_in)
}
\arguments{
\item{dat_in}{is list of data frames}
}
\value{
A \code{\link[base]{data.frame}} of the range of values for each variable
}
\description{
Get the range of values for the same water quality variable between dates
}
