% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PB_dep_pts.R
\docType{data}
\name{PB_dep_pts}
\alias{PB_dep_pts}
\title{Subsample of Pensacola Bay depth soundings, from http://estuarinebathymetry.noaa.gov/bathy_htmls/G130.html}
\format{
A \code{\link[base]{data.frame}} with 2500 rows and 3 variables:
\describe{
  \item{\code{Depth}}{numeric, depth at the location, negative is on land}
  \item{\code{Lat}}{numeric, latitude decimal degrees}
  \item{\code{Long}}{numeric, longitude decimal degrees}
}
}
\usage{
PB_dep_pts
}
\description{
Subsample of Pensacola Bay depth soundings, from http://estuarinebathymetry.noaa.gov/bathy_htmls/G130.html
}
\examples{
data(PB_dep_pts)
}
\keyword{datasets}
