% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PB_pts.R
\docType{data}
\name{PB_pts}
\alias{PB_pts}
\title{Lat/Long locations of Pensacola Bay monthly CTD stations}
\format{A \code{\link[base]{data.frame}} with 9 rows and 3 variables:
\describe{
  \item{\code{Station}}{chr, station name}
  \item{\code{Lat}}{numeric, latitude decimal degrees}
  \item{\code{Long}}{numeric, longitude decimal degrees}
}}
\usage{
PB_pts
}
\description{
Lat/Long locations of Pensacola Bay monthly CTD stations
}
\examples{
data(PB_pts)
}
\keyword{datasets}

