% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ctd_plot.R, R/ctd_plotmult.R
\name{rotate}
\alias{rotate}
\title{Matrix rotation}
\usage{
rotate(x)

rotate(x)
}
\arguments{
\item{x}{input matrix}

\item{x}{input matrix}
}
\description{
Rotate a matrix

Rotate a matrix
}
\details{
Used internally in \code{\link{ctd_plot}}

Used internally in \code{\link{ctd_plot}}
}

