% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ctd_plot.R
\name{ctd_plot}
\alias{ctd_plot}
\title{Contour plot for CTD data}
\usage{
ctd_plot(dat_in, var_plo, rngs_in = NULL, num_levs = 8,
  ylab = "Depth (m)", xlab = "Channel distance from P01 to P09 (km)",
  var_lab = NULL, cols = c("tomato", "lightblue", "lightgreen", "green"),
  ncol = 100)
}
\arguments{
\item{dat_in}{ctd data along the tidal axis, as a \code{\link[base]{data.frame}}}

\item{var_plo}{chr string of variable to plot from \code{dat_in}}

\item{rngs_in}{output from \code{\link{get_rngs}}, used to scale the colors, see examples}

\item{num_levs}{numeric for number of contour levels}

\item{ylab}{chr string for y-axis label}

\item{xlab}{chr string for x-axis label}

\item{var_lab}{optional chr string of text to include in the plot}

\item{cols}{chr string of colors for the plot}

\item{ncol}{numeric indicating degree of smoothing for the color palette}
}
\value{
The contour plot
}
\description{
Plot two-dimensional contours for CTD data by depth and distance
}
\details{
Raw data from the CTD are vertical profiles at unique stations.  The data are linearly interpolated along each profile and between stations to create a two-dimensional plotting surface.  The function currently only works for Pensacola Bay using hard-coded depth transect data.
}
\examples{
# default plot
ctd_plot(ctd_ex1, 'Salinity')

# make color ramp match contour categories in legend
ctd_plot(ctd_ex1, 'Salinity', ncol = 8)

# change colors
ctd_plot(ctd_ex1, 'Salinity', cols = c('Blue', 'Purple', 'Orange'))
}
\references{
Hagy III, JD, Murrell, MC. 2007. Susceptibility of a northern Gulf of Mexico estuary to hypoxia: An analysis using box models. Estuarine, Coastal, and Shelf Science. 74:239-253.
}

